[t] Adding tests of blessed hashes
This is a corollary to the issue mentioned in RT#75561.  It turns out that
the XS code can't determine if the input is a blessed hash (in the case of
the ticket, a blessed array) and thus returns the blessed hash (array)
instead of undef.  This test exposes this issue.