Allowing POSINT to accept regexp matches as args
This bug was actually very subtle.  A POSINT is something which is (a)
defined, (b) not a reference and (c) matches a regexp for a positive
integer.  If these evaluate as true, then the argument is returned,
otherwise undef is returned.  As it turns out, when a regexp match is passed
in as an argument to this function, then (a), (b) and (c) all evaluate to
true, however undef was *still* returned.  The reason being that (c) must
use $_[0] and sets it to undef, which was then being returned.  The fix is
thus to set the argument to POSINT ($_[0]) to a variable and then use this
variable in (a), (b) and (c) to process the function correctly.  It might
possibly be a good idea to use 'shift'  to obtain the input argument, but
that can be improved as needed later.

This change should address the issue raised in RT#81276.